Move cocoa backend to separate file
